Add a test for server checkout and CheckoutCommand (#325)
