+++ plugin/src/com/microsoft/alm/plugin/idea/tfvc/core/TfvcEditFileProvider.java
@@ -7,8 +7,8 @@ import com.intellij.openapi.vcs.EditFileProvider;
 import com.intellij.openapi.vcs.VcsException;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.microsoft.alm.plugin.context.ServerContext;
+import com.microsoft.alm.plugin.external.utils.TfvcCheckoutResultUtils;
 import com.microsoft.alm.plugin.idea.common.resources.TfPluginBundle;
-import com.microsoft.tfs.model.connector.TfsLocalPath;
 import com.microsoft.tfs.model.connector.TfvcCheckoutResult;
 
 import java.nio.file.Path;
@@ -34,19 +34,7 @@ public class TfvcEditFileProvider implements EditFileProvider {
                 serverContext,
                 paths,
                 false);
-        if (!result.getErrorMessages().isEmpty()) {
-            String message = String.join("\n", result.getErrorMessages());
-            throw new VcsException(TfPluginBundle.message(TfPluginBundle.KEY_TFVC_CHECKOUT_FAILED, message));
-        }
-
-        if (!result.getNotFoundFiles().isEmpty()) {
-            List<String> failedPaths = result.getNotFoundFiles().stream()
-                    .map(TfsLocalPath::getPath)
-                    .collect(Collectors.toList());
-            String failedPathList = String.join(", ", failedPaths);
-            throw new VcsException(
-                    TfPluginBundle.message(TfPluginBundle.KEY_TFVC_CHECKOUT_FILES_FAILED, failedPathList));
-        }
+        TfvcCheckoutResultUtils.verify(result);
     }
 
     @Override
